Refactor flic
